Return active storage Blob ID to `bigint`

+++ db/schema.rb
@@ -45,7 +45,7 @@ ActiveRecord::Schema.define(version: 2022_11_29_222608) do
   end
 
   create_table "active_storage_variant_records", force: :cascade do |t|
-    t.integer "blob_id", null: false
+    t.bigint "blob_id", null: false
     t.string "variation_digest", null: false
     t.index ["blob_id", "variation_digest"], name: "index_active_storage_variant_records_uniqueness", unique: true
   end
